feat(dapp): remove unused functions and rename hook
